pyproject.toml
[tool.poetry]
name = "parser"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
Django = "^5.2.3"
django-model-utils = "^5.0.0"
django-modeltranslation = "^0.19.12"
pytest = "^8.3.4"
pytest-django = "^4.10.0"
factory-boy = "^3.3.3"
faker = "^36.1.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
django-mptt = "^0.17.0"
pillow = "^11.1.0"
aiohttp = "^3.11.12"
requests = "^2.32.3"
celery = "^5.4.0"
redis = "^5.2.1"
rapidfuzz = "^3.12.1"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
django-filter = "^25.1"
django-redis = "^5.4.0"
django-cors-headers = "^4.7.0"
gunicorn = "^23.0.0"
django-celery-beat = "^2.7.0"
pytz = "^2025.1"
model-bakery = "^1.20.5"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "parser.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-ra -q"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
